+++ apps/web/modules/shell/banners/LayoutBanner.tsx
@@ -11,10 +11,12 @@ import {
 import VerifyEmailBanner, {
   type VerifyEmailBannerProps,
 } from "@calcom/web/modules/users/components/VerifyEmailBanner";
+import { DentalTwoFactorBanner } from "@calcom/web/modules/settings/security/DentalTwoFactorBanner";
 
 type BannerTypeProps = {
   verifyEmailBanner: VerifyEmailBannerProps;
   adminPasswordBanner: AdminPasswordBannerProps;
+  dentalTwoFactorBanner: boolean | null;
   calendarCredentialBanner: CalendarCredentialBannerProps;
   invalidAppCredentialBanners: InvalidAppCredentialBannersProps;
 };
@@ -30,6 +32,7 @@ export type AllBannerProps = { [Key in BannerType]: BannerTypeProps[Key]["data"]
 export const BannerComponent: BannerComponent = {
   verifyEmailBanner: (props: VerifyEmailBannerProps) => <VerifyEmailBanner {...props} />,
   adminPasswordBanner: (props: AdminPasswordBannerProps) => <AdminPasswordBanner {...props} />,
+  dentalTwoFactorBanner: (_props: { data: boolean | null }) => <DentalTwoFactorBanner />,
   calendarCredentialBanner: (props: CalendarCredentialBannerProps) => <CalendarCredentialBanner {...props} />,
   invalidAppCredentialBanners: (props: InvalidAppCredentialBannersProps) => (
     <InvalidAppCredentialBanners {...props} />
@@ -48,6 +51,10 @@ export const BannerContainer: React.FC<BannerContainerProps> = ({ banners }) => 
           const Banner = BannerComponent[key];
           return <Banner data={banners[key]} key={key} />;
         } else if (key === "adminPasswordBanner") {
+          const Banner = BannerComponent[key];
+          return <Banner data={banners[key]} key={key} />;
+        } else if (key === "dentalTwoFactorBanner") {
+          if (!banners[key]) return null;
           const Banner = BannerComponent[key];
           return <Banner data={banners[key]} key={key} />;
         } else if (key === "calendarCredentialBanner") {